fix: асинхронная версия

+++ routes/edit_form/process_edit_form.py
@@ -2,28 +2,20 @@ from lib.all_configs import read_config
 from lib.core import exists_arg, random_filename
 
 import re 
-def form_update_or_insert(form):
+async def form_update_or_insert(form):
+
     if form.read_only:
       form.errors.append('Вам запрещено сохранять изменения')
     else:
       if form.action=='update':
-          form.run_event('before_update')
+          await form.run_event('before_update')
       if form.action=='insert':
-          form.run_event('before_insert')
+          await form.run_event('before_insert')
       if form.action in ('insert','update'):
-          form.run_event('before_save')
+          await form.run_event('before_save')
       
       if not len(form.errors):
-        form.save()
-      
-        if form.action=='update':
-          form.run_event('after_update')
-
-        if form.action=='insert':
-          form.run_event('after_insert')
-
-        if form.action in ['insert','update']:
-          form.run_event('after_save')
+        await form.save()
 
     return {
       'success':form.success(),
@@ -32,7 +24,7 @@ def form_update_or_insert(form):
       'id':form.id
     }
 
-def process_edit_form(**arg):
+async def process_edit_form(**arg):
   action=arg['action']
   config=arg['config']
   R=arg['R']
@@ -42,7 +34,8 @@ def process_edit_form(**arg):
     values=R['values']
   if 'id' not in arg: arg['id']=''
   
-  form=read_config(
+  form = await read_config(
+    request=arg['request'],
     action=action,
     config=config,
     id=arg['id'],
@@ -50,55 +43,46 @@ def process_edit_form(**arg):
     values=values,
     script='edit_form'
   )
+  if hasattr(form, 'response') and form.response:
+    return form.response
+  
+  if len(form.errors): return {'success':False,'errors':form.errors}
 
   need_fields=[]
   
   for f in form.fields:
+
     if not('orig_type' in f) or not re.search(r'^filter_extend_',f['orig_type']):
       need_fields.append(f)
 
   form.fields=need_fields
-  
 
-  field=None
-  if 'name' in R and R['name']:
-    field=form.fields_hash[R['name']]
-  
-  #print("\nField:\n",field)
-  #if len(form.errors): return form
-  
   if form.not_create and form.action in ['insert','new']:
     form.read_only=1
   
   if form.action == 'insert' and form.not_create:
     form.errors.append('Вам запрещено создавать новые записи')
   
-  #form.set_orig_types() # уже есть в all_configs.read_config
-  
-  # Перенёс в .lib.all_configs.read_form
-  #form.run_event('permissions')
-  
-  #form.get_values()
-  #form.run_all_before_code()
+
   
   if form.action in ['update','insert']:
     form.new_values=values
     form.check() # проверяем все поля в new_values
-    return form_update_or_insert(form)
+    return await form_update_or_insert(form)
   
   elif form.action == 'delete_file':
-    return form.DeleteFile()
+    return await form.DeleteFile()
   
   elif form.action == 'upload_file':
-    return form.UploadFile()
+    return await form.UploadFile()
     
   else:
     if form.action in ['new','edit']:
       if len(form.errors): form.read_only=1
     
     
-    form.edit_form_process_fields()
-    #form.pre(form.fields[0]['value'])
+    await form.edit_form_process_fields()
+
     return {
       'action':form.action,
       'title':form.title,
@@ -108,9 +92,11 @@ def process_edit_form(**arg):
       'id':form.id,
       'log':form.log,
       'read_only':form.read_only,
-      'width':form.width,
+      'wide_form':(hasattr(form, 'wide_form') and form.wide_form),
       'cols':form.cols,
       'tabs':form.tabs,
       'config':form.config,
-      'javascript':exists_arg('edit_form',form.javascript)
+      'javascript':exists_arg('edit_form',form.javascript),
+      'javascript_static':exists_arg('edit_form_static',form.javascript),
+      'redirect':(hasattr(form, 'redirect') and form.redirect),
     }
